Refactor graph.rs to be fully mathematically parameterizable

+++ src/client/backend/exercise/weight.rs
@@ -1,6 +1,6 @@
 use std::cmp::Ordering;
 
-pub type Kg = f64;
+pub type Kg = f32;
 pub enum ExerciseWeight {
     BodyweightPlusKg(Kg),
     Kg(Kg),
